+++ agent/src/common/meta_packet.rs
@@ -157,7 +157,7 @@ pub struct MetaPacket<'a> {
     pub is_response_end: bool,
 
     pub process_id: u32,
-    pub netns_id: u64,
+    pub netns_id: u32,
 
     pub thread_id: u32,
     pub syscall_trace_id: u64,
