+++ agent/src/metric/document.rs
@@ -314,7 +314,7 @@ pub struct Tagger {
     pub otel_instance: Option<String>,
     pub endpoint: Option<String>,
     pub signal_source: SignalSource,
-    pub netns_id: u64,
+    pub netns_id: u32,
 }
 
 impl Default for Tagger {
@@ -406,7 +406,7 @@ impl From<Tagger> for metric::MiniTag {
                 app_service: t.otel_service.unwrap_or_default(),
                 app_instance: t.otel_instance.unwrap_or_default(),
                 endpoint: t.endpoint.unwrap_or_default(),
-                netns_id: t.netns_id as u32,
+                netns_id: t.netns_id,
             }),
         }
     }
